recomendacion/requirements.txt
flask
pandas
scikit-learn
requests
python-dotenv
